requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiofiles==23.2.0
websockets==12.0
requests==2.31.0

# AI & Vector Store
chromadb==0.4.15
openai==1.3.7
sentence-transformers==2.2.2

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.1.3
numpy==1.25.2
spacy==3.7.2
nltk==3.8.1
beautifulsoup4==4.12.2
openpyxl==3.1.2

# Image Processing & OCR
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
easyocr==1.7.0

# Audio Processing
openai-whisper==20231117
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Video Processing
moviepy==1.0.3
ffmpeg-python==0.2.0

# Background Processing
celery[redis]==5.3.4
redis==5.0.1

# File Type Detection
python-magic==0.4.27
mutagen==1.47.0